Organise all sections into Section folder under component
