+++ components/Sections/SuperpowerBaselineSection/BaselineSlideVertical.tsx
